requirements.txt
# Required python packages for fandjango
# Can be installed py "pip install -r requirements.txt"

Django >= 1.9
django-autoslug
Jinja2
psycopg2
django-ssl-admin

# Use gunicorn in production
gunicorn
setproctitle

# Fetch covers
beautifulsoup4
Pillow
